Move DisplayVariables into frontend directory
